+++ src/public/img/blue-check.png
