+++ src/public/js/emoji_generator.js
