+++ src/public/js/language_colors_generator.js
