+++ src/public/js/spinner.js
